tsk 21 Fixes Story test
The test class was moved to the Domain project from the Tests project as this
is only for utility classes used in the tests.
The class was renamed to StoryEntityTest as all test classes must use the
Test suffix.
Method testStoryLengthStory has been removed as it was testing some of the same
functionality as the testStoryDataType method.
Further tests were added to check the content size for each type of genre as
each genre has a different size limit.