tsk 10 Adds service functionality to get stories containing a text
A new method has been added to the StoryEJB class to get the stories that
contain a given text in their title or content.
This method paginates the results and retrieves 9 stories per page.
Two tests has been created for this method. One looks for stories containing
a given text and the other looks for a text that doesn't exist.