Fixes sql and modifies jsf
Adds the User table to the sql and its foreing key to map to its entity.
Fixes the relationship of the user entity with story so that it is displays in jsf.
The stories are displays in full page and not in a modal,
this is to avoids bugs when creates the visits.