tsk-26 Creates resource method to get the most read stories
Created a new resource method to retrieve the most read stories between two
dates.
The dates passed as parameters must be Strings with the following format:
yyyy-MM-dd (2022-05-25).
Also, the list of stories can be filtered by genre and it is possible to
indicate the page and maximum items to be returned.
The method was successfully tested using Postman.