tsk 28 Adds login and logout functionality
Created loginController for the login and logout. Both login and logout
functions use the HttpSession class to set the user session.
Author and Admin roles were created to control access to the pages.
A function in the UserEJB was created to validate the credentials of the user
trying to log in.
A common page for the navigation was created to avoid code repetition in each
page. All pages have been updated to import this common navigation.